backend/requirements.txt
alembic==1.16.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
pandas==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.42
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
PyJWT
Werkzeug
pytest
pytest-cov
